+++ plugin/brdrq/brdrq.py
@@ -34,7 +34,9 @@ import inspect
 import os
 import sys
 
+import brdr
 import numpy as np
+from PyQt5.QtCore import pyqtSignal
 from PyQt5.QtGui import QIcon
 from PyQt5.QtWidgets import QAction, QListWidgetItem
 from brdr.constants import PREDICTION_SCORE, EVALUATION_FIELD_NAME
@@ -46,6 +48,7 @@ from qgis.core import QgsMapLayerProxyModel
 from qgis.core import QgsProject
 from qgis.core import QgsSettings
 from qgis.core import QgsStyle
+from qgis.gui import QgsMapToolPan
 from qgis.utils import OverrideCursor
 from shapely.io import from_wkt
 
@@ -91,6 +94,8 @@ class BrdrQPlugin(object):
         self.actions = []
         self.toolbar = self.iface.addToolBar("brdrQ")
         self.toolbar.setObjectName("brdrQ")
+        self.selectTool = None
+        self.formerMapTool = None
         self.max_rel_dist = None
         self.minimum = 0
         self.maximum = 1500
@@ -102,7 +107,7 @@ class BrdrQPlugin(object):
             400000  # maximum m² where the calculation will be done for
         )
         self.layer = None
-        self.selected_features = None
+        self.listed_features = None
         self.feature = None
         self.dict_processresults = None
         self.dict_evaluated_predictions = None
@@ -121,6 +126,7 @@ class BrdrQPlugin(object):
         self.partial_snapping = None
         self.partial_snapping_strategy = None
         self.snap_max_segment_length = None
+        self.DECIMAL = 1
         self.GROUP_LAYER = "brdrQ_plugin"
         self.LAYER_RESULT = (
             "RESULT"  # parameter that holds the TOC layername of the result
@@ -158,9 +164,9 @@ class BrdrQPlugin(object):
     def initGui(self):
         # print ("initGui")
         self.initProcessing()
-        icon = os.path.join(os.path.join(cmd_folder, "icon.png"))
+        icon = os.path.join(os.path.join(cmd_folder, "icon_featurealigner.png"))
         action_featurepredictor = QAction(
-            QIcon(icon), "brdrQ - GRB actual Parcel Aligner", self.iface.mainWindow()
+            QIcon(icon), "brdrQ - Feature Aligner (predictor)", self.iface.mainWindow()
         )
         action_featurepredictor.triggered.connect(self.openDock)
         self.iface.addPluginToMenu("brdQ", action_featurepredictor)
@@ -171,14 +177,39 @@ class BrdrQPlugin(object):
         )
         action_autocorrectborders = QAction(
             QIcon(icon_autocorrectborders),
-            "Autocorrectborders",
+            "Autocorrectborders (bulk)",
             self.iface.mainWindow(),
         )
         action_autocorrectborders.triggered.connect(self.openAutocorrectbordersscript)
-        # self.iface.addToolBarIcon(action)
         self.iface.addPluginToMenu("brdQ", action_autocorrectborders)
         self.toolbar.addAction(action_autocorrectborders)
         self.actions.append(action_autocorrectborders)
+
+        icon_autoupdateborders = os.path.join(
+            os.path.join(cmd_folder, "icon_autoupdateborders.png")
+        )
+        action_autoupdateborders = QAction(
+            QIcon(icon_autoupdateborders),
+            "Autoupdateborders (bulk)",
+            self.iface.mainWindow(),
+        )
+        action_autoupdateborders.triggered.connect(self.openAutoupdatebordersscript)
+        self.iface.addPluginToMenu("brdQ", action_autoupdateborders)
+        self.toolbar.addAction(action_autoupdateborders)
+        self.actions.append(action_autoupdateborders)
+
+        icon_info = os.path.join(
+            os.path.join(cmd_folder, "icon_info.png")
+        )
+        action_info = QAction(
+            QIcon(icon_info),
+            "brdrQ - info (version)",
+            self.iface.mainWindow(),
+        )
+        action_info.triggered.connect(self.openInfo)
+        self.iface.addPluginToMenu("brdQ", action_info)
+        self.toolbar.addAction(action_info)
+        self.actions.append(action_info)
 
         # show the dockwidget
         # self.openDock()
@@ -226,6 +257,10 @@ class BrdrQPlugin(object):
             self.settingsDialog.spinBox_snap_max_segment_length.setEnabled(False)
         return
 
+    # def update_checkBox_only_selected(self, state):
+    #     self.listFeatures()
+    #     return
+
     def updateFields_reference(self):
         layer = self.settingsDialog.mMapLayerComboBox_reference.currentLayer()
         self.settingsDialog.mFieldComboBox_reference.setLayer(layer)
@@ -260,7 +295,7 @@ class BrdrQPlugin(object):
         self.maximum = self.max_rel_dist * 100
 
         self.relevant_distances = [
-            round(k, 1)
+            round(k, self.DECIMAL)
             for k in np.arange(
                 self.minimum, self.maximum + self.step, self.step, dtype=int
             )
@@ -270,7 +305,7 @@ class BrdrQPlugin(object):
             self.dockwidget.doubleSpinBox.setMinimum(self.minimum / 100)
             self.dockwidget.doubleSpinBox.setMaximum(self.maximum / 100)
             self.dockwidget.doubleSpinBox.setSingleStep(self.step / 100)
-            self.dockwidget.doubleSpinBox.setDecimals(1)
+            self.dockwidget.doubleSpinBox.setDecimals(self.DECIMAL)
             self.dockwidget.doubleSpinBox.setValue(0.0)
             self.dockwidget.horizontalSlider.setMinimum(0)
             self.dockwidget.horizontalSlider.setMaximum(
@@ -403,6 +438,21 @@ class BrdrQPlugin(object):
         )
         dialog_autocorrectborders.exec()
 
+    def openInfo(self):
+        msg = f"brdrQ version: {self.version()} - brdr-version: {str(brdr.__version__)}"
+        self.iface.messageBar().pushMessage(msg)
+
+    def version(self):
+        return "0.9.12"
+
+
+    def openAutoupdatebordersscript(self):
+        dialog_autoupdateborders = processing.createAlgorithmDialog(
+            "brdrqprovider:brdrqautoupdateborders"
+        )
+        dialog_autoupdateborders.exec()
+
+
     def remove_brdrq_layers(self):
         tree = QgsProject.instance().layerTreeRoot()
         node_object = tree.findGroup(self.GROUP_LAYER)
@@ -461,12 +511,15 @@ class BrdrQPlugin(object):
             )
             self.dockwidget.pushButton_save.clicked.connect(self.change_geometry)
             self.dockwidget.pushButton_reset.clicked.connect(self.reset_geometry)
-            # self.dockwidget.pushButton_select.clicked.connect(self.start_line_edit)
-            self.dockwidget.pushButton_wkt.clicked.connect(self.get_wkt)
+            self.dockwidget.pushButton_select.clicked.connect(self.activate_selectTool)
+            #self.dockwidget.pushButton_wkt.clicked.connect(self.get_wkt)
             self.dockwidget.mMapLayerComboBox.setFilters(
                 QgsMapLayerProxyModel.PolygonLayer
             )
-            self.dockwidget.mMapLayerComboBox.layerChanged.connect(self.setFeatures)
+            self.dockwidget.mMapLayerComboBox.layerChanged.connect(self.themeLayerChanged)
+            self.dockwidget.checkBox_only_selected.stateChanged.connect(
+                self.themeLayerChanged
+            )
             self.dockwidget.listWidget_features.itemPressed.connect(
                 self.onFeatureActivated
             )
@@ -494,8 +547,10 @@ class BrdrQPlugin(object):
     #     picker.setLayer(layer)
     #     picker.setDisplayExpression('$id')  # show ids in combobox
 
-    def setFeatures(self):
-        self.dockwidget.progressBar.setValue(0)
+    def themeLayerChanged(self):
+        #self.deactivateSelectTool()
+        #reset interface by clearing list, progress_bar
+        self._clearUserInterface()
         self.layer = self.dockwidget.mMapLayerComboBox.currentLayer()
         if self.layer is None:
             self.dockwidget.textEdit_output.setText("Please select a layer")
@@ -508,18 +563,42 @@ class BrdrQPlugin(object):
                 f"Nr of features bigger than {str(self.max_feature_count)}. Please make a smaller selection of features"
             )
             return
-        self.selected_features = [f for f in self.layer.getSelectedFeatures()]
-        if self.layer.selectedFeatureCount() == 0:
-            self.selected_features = [f for f in self.layer.getFeatures()]
+        # self.listed_features = [f for f in self.layer.getSelectedFeatures()]
+        # if self.layer.selectedFeatureCount() == 0:
+        #     self.listed_features = [f for f in self.layer.getFeatures()]
+        #     self.dockwidget.textEdit_output.setText(
+        #         "No selected features in this layer, all features returned"
+        #     )
+        # elif self.dockwidget.checkBox_only_selected.checkState()==0:
+        #     self.listed_features = [f for f in self.layer.getFeatures()]
+        #     self.dockwidget.textEdit_output.setText(
+        #         "All features in this layer returned"
+        #     )
+
+        if self.dockwidget.checkBox_only_selected.checkState()==2:
+            self.listed_features = [f for f in self.layer.getSelectedFeatures()]
             self.dockwidget.textEdit_output.setText(
-                "No selected features in this layer, all features returned"
+                "Selected features in this layer returned"
             )
+        else:
+            self.listed_features = [f for f in self.layer.getFeatures()]
+            self.dockwidget.textEdit_output.setText(
+                "All features in this layer returned"
+            )
+        self.listFeatures()
 
-        # Clear the list widget
-        self.dockwidget.listWidget_features.clear()
+    def deactivateSelectTool(self):
+        mapcanvas =self.iface.mapCanvas()
+        if self.formerMapTool is None:
+            self.formerMapTool = QgsMapToolPan(mapcanvas)
+        mapcanvas.setMapTool(self.formerMapTool)
 
+    def listFeatures(self):
+        self._clearUserInterface()
         # Add the selected features to the list widget
-        for feature in self.selected_features:
+        print ("list features")
+        print(str(self.listed_features))
+        for feature in self.listed_features:
             attributes = feature.attributes()
             attribute_string = ", ".join(str(attribute) for attribute in attributes)
             item = QListWidgetItem(
@@ -527,11 +606,27 @@ class BrdrQPlugin(object):
             )
             self.dockwidget.listWidget_features.addItem(item)
         self.dockwidget.textEdit_output.setText(
-            f"#Features: {str(len(self.selected_features))}"
+            f"#Features: {str(len(self.listed_features))}"
         )
+        if len(self.listed_features) == 1:
+            self.onFeatureActivated(self.dockwidget.listWidget_features.item(0))
+
         return
 
+    def _clearUserInterface(self):
+        # Clear progressbar
+        self.dockwidget.progressBar.setValue(0)
+        self.dockwidget.doubleSpinBox.setValue(0)
+        # Clear the featurelist widget
+        self.dockwidget.listWidget_features.clear()
+        # Clear the predictionlist
+        self.dockwidget.listWidget_predictions.clear()
+        self.dockwidget.checkBox_only_selected.setEnabled(True)
+        # if self.layer is not None and self.layer.selectedFeatureCount() == 0:
+        #     self.dockwidget.checkBox_only_selected.setEnabled(False)
+
     def onFeatureActivated(self, currentItem):
+        self.deactivateSelectTool()
         self.dockwidget.progressBar.setValue(0)
         self.dockwidget.listWidget_predictions.clear()
         self.dockwidget.textEdit_output.setText("")
@@ -546,14 +641,10 @@ class BrdrQPlugin(object):
             print("currentItem is none")
             return
         feature_id = currentItem.text().split("*")[1]
-        # print(f"Feature_id is {feature_id}")
-        for feat in self.selected_features:
-            # print(str(feat.id()))
+        for feat in self.listed_features:
             if str(feat.id()) == feature_id:
-                # print("equal - >break")
                 self.feature = feat
                 break
-        # print(self.feature)
         if self.feature is None:
             self.dockwidget.textEdit_output.setText(
                 f"No feature found with ID {feature_id}"
@@ -575,12 +666,12 @@ class BrdrQPlugin(object):
             if area > self.max_area_limit:
                 msg = f"Very big area, {str(area)} m²: The calculation is blocked. Please use the bulk tool for this feature"
                 self.dockwidget.textEdit_output.setText(f"{msg}")
-                # self.iface.messageBar().pushMessage(msg)
+                self.dockwidget.doubleSpinBox.setValue(0)
+                self.dockwidget.listWidget_predictions.clear()
                 return
             else:
                 msg = f"Warning - Big area, {str(area)} m²: the calculation will be adapted/optimized. Only for every meter a calculation will be done"
                 self.dockwidget.textEdit_output.setText(f"{msg}")
-                # self.iface.messageBar().pushMessage(msg)
                 self.step = 100
 
         self._update_settings()
@@ -650,12 +741,14 @@ class BrdrQPlugin(object):
         self.dockwidget.listWidget_predictions.addItems(items_with_name)
         if len(items) > 0:
             self.dockwidget.listWidget_predictions.setCurrentRow(best_index)
-            self.dockwidget.doubleSpinBox.setValue(round(float(items[best_index]), 1))
+            print ("best-index: "+str(items[best_index]))
+            self.dockwidget.doubleSpinBox.setValue(round(float(items[best_index]), self.DECIMAL))
         else:
             self.dockwidget.textEdit_output.setText("No predictions")
 
     def onListItemActivated(self, currentItem):
         print("onListItemActivated")
+        self.deactivateSelectTool()
         self._listItemActivated(currentItem)
 
     def _listItemActivated(self, currentItem):
@@ -666,7 +759,7 @@ class BrdrQPlugin(object):
         print("item activated with rd: " + currentItem.text())
         value = currentItem.text()
         value = value.split(":")[0]
-        value = round(float(value), 1)
+        value = round(float(value), self.DECIMAL)
         print("item activated with rd - value: " + str(value))
         self.dockwidget.doubleSpinBox.setValue(value)
         index = self.relevant_distances.index(value)
@@ -676,11 +769,12 @@ class BrdrQPlugin(object):
     def onSliderChange(self, index):
         print("onSliderChange: index -> " + str(index))
         value = self.relevant_distances[index]
+        value = round(value, self.DECIMAL)
         self.dockwidget.doubleSpinBox.setValue(value)
         return
 
     def onSpinboxChange(self, value):
-        value = round(value, 1)
+        value = round(value, self.DECIMAL)
         index = self.relevant_distances.index(value)
         self.dockwidget.horizontalSlider.setValue(index)
         print("onSpinboxChange: value -> " + str(value))
@@ -699,12 +793,32 @@ class BrdrQPlugin(object):
         get_layer_by_name(self.LAYER_RESULT_DIFF_PLUS).setSubsetString(
             f"brdr_relevant_distance = {value}"
         )
+        self.get_wkt()
         return
 
-    # def start_line_edit(self):
-    #     print("button pushed")
-    #     self.t = selectTool(self.iface, self.dockwidget.mMapLayerComboBox.currentLayer())
-    #     self.iface.mapCanvas().setMapTool(self.t)
+    def activate_selectTool(self):
+        print("button pushed")
+        print ("currentlayer:" + str (self.dockwidget.mMapLayerComboBox.currentLayer()))
+        self.selectTool = SelectTool(self.iface, self.dockwidget.mMapLayerComboBox.currentLayer())
+        self.formerMapTool = self.iface.mapCanvas().mapTool()
+        self.iface.mapCanvas().setMapTool(self.selectTool)
+        #self.selectTool.featureIdentified.connect(self.callback_select)
+        self.selectTool.featuresIdentified.connect(self.onFeaturesIdentified)
+        print("end activate_selecttool")
+
+    def onFeaturesIdentified(self,identified_features):
+        """Code called when the feature is selected by the user"""
+        self.listed_features = identified_features
+        self.listFeatures()
+        # if len(identified_features)>0:
+        #     #self.layer.removeSelection()
+        #     #self.layer.selectByIds([f.id() for f in identified_features], QgsVectorLayer.AddToSelection)
+        #     self.listed_features = identified_features
+        #     self.listFeatures()
+        #     #self.layer.removeSelection()
+        # else:
+        #     msg =  (f"no features selected in layer {self.layer.name()}")
+        #     self.dockwidget.textEdit_output.setText(f"{msg}")
 
     def get_graphic(self):
         # feat = self.dockwidget.mFeaturePickerWidget.feature()
@@ -735,7 +849,7 @@ class BrdrQPlugin(object):
         if feat is None:
             return
         key = feat.id()
-        relevant_distance = self.dockwidget.doubleSpinBox.value()
+        relevant_distance = round(self.dockwidget.doubleSpinBox.value(),self.DECIMAL)
         if relevant_distance in self.dict_processresults[key]:
             result = self.dict_processresults[key][relevant_distance]
             resulting_geom = result["result"]
@@ -743,7 +857,6 @@ class BrdrQPlugin(object):
             errormesssage = "Relevant_distance_result not calculated for: " + str(
                 relevant_distance
             )
-            self.iface.messageBar().pushMessage(errormesssage)
             print(errormesssage)
             return
         # layer = self.dockwidget.mMapLayerComboBox.currentLayer()
@@ -772,15 +885,25 @@ class BrdrQPlugin(object):
         if feat is None:
             return
         key = feat.id()
-        relevant_distance = self.dockwidget.doubleSpinBox.value()
-        if relevant_distance in self.dict_processresults[key]:
+        print ("key:" + str(key))
+        relevant_distance = round(self.dockwidget.doubleSpinBox.value(),self.DECIMAL)
+        print (str(relevant_distance))
+        if key is None or self.dict_processresults is None or not key in self.dict_processresults.keys():
+            msg = f"No prediction-WKT of feature {str(key)}..."
+            self.dockwidget.textEdit_output.setText(msg)
+            return
+
+
+        elif relevant_distance in self.dict_processresults[key]:
             result = self.dict_processresults[key][relevant_distance]
             resulting_geom = result["result"]
         else:
-            errormesssage = "Relevant_distance_result not calculated for: " + str(
+            errormesssage = "Relevant_distance_result not calculated for key : " + str(
+                key
+            )+ " at relevant distance-" + str(
                 relevant_distance
             )
-            self.iface.messageBar().pushMessage(errormesssage)
+
             self.dockwidget.textEdit_output.setText(errormesssage)
             return
         wkt = resulting_geom.wkt
@@ -850,13 +973,7 @@ class BrdrQPlugin(object):
             self.aligner.dict_reference_source["source"] = "local"
             self.aligner.dict_reference_source["version_date"] = "unknown"
         self.dockwidget.progressBar.setValue(50)
-        # self.dict_series, self.dict_predictions, self.diffs_dict = (
-        #     self.aligner.predictor(
-        #         relevant_distances=self.relevant_distances,
-        #         od_strategy=self.od_strategy,
-        #         threshold_overlap_percentage=self.threshold_overlap_percentage,
-        #     )
-        # )
+
         dict_evaluated, props_dict_evaluated_predictions = self.aligner.evaluate(
             ids_to_evaluate=None,
             base_formula_field=None,
@@ -876,8 +993,6 @@ class BrdrQPlugin(object):
             [str(k) for k in self.dict_evaluated_predictions[feat.id()].keys()]
         )
         self.dockwidget.textEdit_output.setText(outputMessage)
-        self.iface.messageBar().pushMessage(outputMessage)
-        # self.dockwidget.progressBar.setValue(100)
         return (
             self.dict_processresults,
             self.dict_evaluated_predictions,
@@ -885,29 +1000,36 @@ class BrdrQPlugin(object):
         )
 
 
-# from qgis.gui import QgsMapToolIdentifyFeature, QgsMapToolIdentify
-# from qgis.core import (
-#     Qgis, QgsVectorLayer
-# )
-# class selectTool(QgsMapToolIdentifyFeature):
-#
-#     def __init__(self, iface, layer):
-#         self.iface = iface
-#         self.canvas = self.iface.mapCanvas()
-#         self.layer = layer
-#         QgsMapToolIdentifyFeature.__init__(self, self.canvas, self.layer)
-#         self.iface.currentLayerChanged.connect(self.active_changed)
-#
-#     def active_changed(self, layer):
-#         self.layer.removeSelection()
-#         if isinstance(layer, QgsVectorLayer) and layer.isSpatial():
-#             self.layer = layer
-#             self.setLayer(self.layer)
-#
-#     def canvasPressEvent(self, event):
-#         found_features = self.identify(event.x(), event.y(), [self.layer], QgsMapToolIdentify.TopDownAll)
-#         print(found_features)
-#         self.layer.selectByIds([f.mFeature.id() for f in found_features], QgsVectorLayer.AddToSelection)
-#
-#     def deactivate(self):
-#         self.layer.removeSelection()
+from qgis.gui import QgsMapToolIdentifyFeature, QgsMapToolIdentify
+from qgis.core import (
+    QgsVectorLayer
+)
+class SelectTool(QgsMapToolIdentifyFeature):
+    featuresIdentified = pyqtSignal(object)
+    def __init__(self, iface,layer):
+        self.iface = iface
+        self.canvas = self.iface.mapCanvas()
+        self.layer = layer
+        # self.plugin = brdrQPlugin
+        QgsMapToolIdentifyFeature.__init__(self, self.canvas, self.layer)
+        self.iface.currentLayerChanged.connect(self.active_changed)
+
+
+    def active_changed(self, layer):
+        print ("active_changed")
+        #self.layer.removeSelection()
+        if isinstance(layer, QgsVectorLayer) and layer.isSpatial():
+            self.layer = layer
+            self.setLayer(self.layer)
+
+    def canvasPressEvent(self, event):
+        identified_features = self.identify(event.x(), event.y(), [self.layer], QgsMapToolIdentify.TopDownAll)
+        identified_features = [f.mFeature for f in identified_features]
+        self.featuresIdentified.emit(identified_features)
+        #self.deactivate()
+
+    def deactivate(self):
+        #self.layer.removeSelection()
+        #QgsMapTool.deactivate(self)
+        print("deactivate")
+        #self.iface.mapCanvas().setMapTool(None)